Push getting the keyfile path down one layer
I'ld prefer it if that part of the computation wasn't in the D-Bus layer
now.

The purpose of the existing code is to enable the future possibility of
using a different keyfile for the cache layer than for the data layer. But
when that happens the keyfile will be extracted by the D-Bus layer and
passed in, and at that time we can rearrange init_cache as necessary, both
to take the argument and to use it properly.

Signed-off-by: mulhern <amulhern@redhat.com>