Drop transitive dependency on time crate
It turns out that it is not required by the chrono crate.

It was made optional in chrono version 0.4.16, see:
https://github.com/chronotope/chrono/pull/478.

Require chrono 0.4.16 in Cargo.toml, as before that, the clock
feature pulled in the time crate, so that specifying no defaults
and precisely "clock" and "std" will still pull in the time crate,
transitively.

Signed-off-by: mulhern <amulhern@redhat.com>